dev: support multiple calls for docs and refactor the algorithm for tools

+++ bin/lib/docs/docs.bash
@@ -1,22 +1,41 @@
 #!/usr/bin/env bash
-# shellcheck source=../core/git.bash # @root
+# shellcheck source=../core/git.bash     # @root
+# shellcheck source=../utils/vendor.bash # @consistent @lock
 
 # Example: run docs npm ci
 # Example: run docs npm i nextra@latest
 #
 # Example: run docs dev|build|start|publish
-# TODO:feat run docs -- build start
+# Example: run docs npm ci -- build -- start
 docs() {
   pushd "$(@root)/docs" >/dev/null || exit 1
   trap 'popd >/dev/null' ERR
 
-  case "${1:-}" in
-  dev) npx --no-install next dev ;;
-  build) npx --no-install next build ;;
-  start) npx --no-install next start ;;
-  publish) rm -rf dist/ && TARGET=static npx --no-install next build ;;
-  *) "${@}" ;;
-  esac
+  local arg args=()
+  for arg in "${@}"; do
+    shift
+
+    if [[ ${#args[@]} -eq 0 ]]; then
+      case "${arg}" in
+      install) if ! @consistent; then npm ci && @lock; fi ;;
+      dev) npx --no-install next dev ;;
+      build) npx --no-install next build ;;
+      start) npx --no-install next start ;;
+      publish) rm -rf dist/ && TARGET=static npx --no-install next build ;;
+      --) continue ;;
+      *) args+=("${arg}") ;;
+      esac
+      continue
+    fi
+
+    if [[ "${arg}" == '--' ]]; then
+      "${args[@]}"
+      args=()
+      continue
+    fi
+    args+=("${arg}")
+  done
+  "${args[@]}"
 
   popd >/dev/null || exit 1
 }
